+++ src/Interfaces/TranslatorInterface.php
@@ -7,8 +7,8 @@ interface TranslatorInterface
     /**
      * Translates a phrase.
      *
-     * @param array       $params parameters to inject into phrase
-     * @param string|null $locale optional locale
+     * @param array       $context parameters to inject into phrase
+     * @param string|null $locale  optional locale
      */
-    public function translate(string $phrase, array $params = [], ?string $locale = null): string;
+    public function translate(string $phrase, array $context = [], ?string $locale = null): string;
 }